// app/_util/spotify.types.ts
import { Image } from './common.type'

export type Track = {
    id: string;
    name: string;
    album: {
        id: string;
        album_type: string;
        release_date: string;
        images: Image[];
    };
    external_urls: {
        spotify: string
    };
}

export type TopTracks =  {
    tracks: Track[]
}

export type SpotifyToken = {
    access_token: string
    token_type: string
    expires_in: number
}